Use React as a peer dependency
This commit sets up the package to use React as a peer dependency.

This is the [recommended way](https://stackoverflow.com/a/30454133) to add React to an external library and how many members of the React core team have it setup in [their packages](https://github.com/gaearon/react-side-effect/blob/master/package.json#L33).

This means that html-to-react will now leverage the version of React that the user has installed, instead of potentially using a different version within the module if the user is not on React v18.  This was causing an increased bundle size when installing `v1.5.1` when compared to `v1.5.0` which this should resolve.